+++ packages/build-scripts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.30
+
+- [fix] jest import
+- [feat] support process.env.EXTRA_PLUGIN_DIR to resolve plugins
+- [feat] support plugin api `cancelTask`
+- [feat] support plugin api `hasMethod`
+- [feat] add hook params of `before.${command}.load`
+
 ## 0.1.29
 
 - [feat] add hook params